==> 2_scrape_players_by_league.ipynb <==
# Scraping de players por liga
Nesta seção serão obtidos todos os jogadores envolvidos em todas as transferências de 2000 a 2023 para cada liga

Informações dos jogadores são salvas no diretório 'data/players/{liga}_players.csv'

O arquivo 'data/players/all_players.csv' contém a concatenação dos players de todas as ligas, removendo jogadores repetidos

Têm o tipo abaixo:

import pandas as pd
pd.read_csv('data/players/all_players.csv')

import requests
from bs4 import BeautifulSoup
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import re
from tqdm import tqdm

headers = {'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/47.0.2526.106 Safari/537.36'}
prefix = 'https://www.transfermarkt.com'
def getSoup(url):
    res = requests.get(url, headers = headers)
    return BeautifulSoup(res.content, 'html.parser')


### Função para realizar o scraping de uma página de jogador

def getPlayerRow(url):
    page = getSoup(prefix + url)
    span_list = page.find_all('span')
    # infoTable = page.find('div', {'class': 'info-table'})
    # span_list = infoTable.find_all('span')

    currAge = 'N/A'
    height = 'N/A'
    foot = 'N/A'
    currClub = 'N/A'

    for i, span in enumerate(span_list):
        if span.text.strip() == 'Age:':
            currAge = span_list[i+1].text.strip()
        if span.text.strip() == 'Height:':
            height = span_list[i + 1].text.replace('m', '').strip().replace(',', '.')
        if span.text.strip() == 'Foot:':
            foot = span_list[i+1].text.strip()
        if span.text.strip() == 'Current club:':
            currClub = span_list[i+1].text.strip()
            
    return ([url, currAge, height, foot, currClub])

def getPlayersTable(playerLinks):
    table = []
    total_iterations = len(playerLinks)
    for link in tqdm(playerLinks, total=total_iterations):
        table.append(getPlayerRow(link))
    df = pd.DataFrame(
        table, columns=['playerLink', 'currAge', 'height', 'foot', 'currClub'])
    df['currAge'] = pd.to_numeric(df['currAge'], errors='coerce').astype(float)
    df['height'] = pd.to_numeric(df['height'], errors='coerce').astype(float)
    return df

pl_transfers = pd.read_csv('data/transfers/pl_transfers.csv')
ll_transfers = pd.read_csv('data/transfers/ll_transfers.csv')
bl_transfers = pd.read_csv('data/transfers/bl_transfers.csv')
sa_transfers = pd.read_csv('data/transfers/sa_transfers.csv')
l1_transfers = pd.read_csv('data/transfers/l1_transfers.csv')

pd.read_csv('data/players/pl_players.csv')

def getPlayersInfo(name, df):
    player_links = df['playerLink'].unique()
    players = getPlayersTable(player_links)
    merged = pd.merge(players, df[['playerLink', 'playerName', 'playerPos' ,'playerCountry', 'playerMarketValue']], on='playerLink', how='left')
    merged = merged.drop_duplicates(subset=['playerLink'])
    players = merged
    players.to_csv('data/players/' + name, index=False)

getPlayersInfo("pl_players.csv", pl_transfers)

getPlayersInfo("ll_players.csv", ll_transfers)

getPlayersInfo("bl_players.csv", bl_transfers)

getPlayersInfo("sa_players.csv", sa_transfers)

getPlayersInfo("l1_players.csv", l1_transfers)

leagues = ["pl", "ll", "bl", "sa", "l1"]
dfs = []
for league in leagues:
    dfs.append(pd.read_csv('data/players/' +league+ '_players.csv'))
all_players = pd.concat(dfs, axis=0)

all_players = all_players.drop_duplicates(subset=['playerLink'], keep='last')

all_players.to_csv('data/players/all_players.csv', index=False)
