==> 4_merge_transfers_stats.ipynb <==
# Merging das tabelas de transferênciasde com stats do fbref para todos os jogadores de todas as ligas
Nesta seção, a tabela de transferências é mergeada com a tabela players, para obter mais informações básicas sobre o jogador, e o resultado é mergeado com as tabelas de stats do fbref, de acordo com o nome do jogador e a temporada de transferência. 

Os arquivos são salvos em 'data/transfers_stats/new' 

Obs.: não estamos usando os old stats por enquanto

Dentro da pasta new, tem informações sobre cada transferência, e as estatísticas do jogador na temporada da transferência
Uma transferência pode ter mais de uma linha, pois o jogador pode ter stats para mais de um clube na temporada.

Para cada liga, foi criada uma pasta na pasta 'data/transfers_stats/new'. Dentro das pastas da liga, foram armazenados os stats das transferências daquela liga

Além disso, para cada tipo de stats, foi criado um arquivo do tipo 'data/transfers_stats/new/all/{type}.csv'

Têm os tipos abaixo:

import pandas as pd

pd.read_csv('data/transfers_stats/new/all/standard.csv')

import pandas as pd
import numpy as np

leagues = ["pl", "ll", "bl", "sa", "l1", "all"]
stats_types = ["standard", "shooting", "passing", "passing_types", "goal_shot_creation", "defense", "possession", "playing_time", "keeper", "misc"]

new_stats = {}
for type in stats_types:
    new_stats[type] = pd.read_csv('data/new_fbref/'+type+'.csv')

def getFbrefFromTransfers(stats, transfers_players, name):
    # renomeando coluna dos stats do fbref para poder mergear depois com a tabela de transferências
    stats = stats.rename(columns={'player': 'playerName'})
    merged_df = pd.merge(transfers_players, stats, on=['playerName', 'season'])
    merged_df.to_csv('data/transfers_stats/'+ name, index=False)

def getTransfersPlayers(league):
    players = pd.read_csv('data/players/'+league+'_players.csv')
    transfers = pd.read_csv('data/transfers/'+ league + '_transfers.csv')
    transfers_players = pd.merge(transfers, players[['playerLink', 'currAge', 'height', 'foot']], on="playerLink", how="left")
    return transfers_players

transfers_players = {}
for league in leagues:
    transfers_players[league] = getTransfersPlayers(league)

for league in leagues:
    for type in stats_types:
        getFbrefFromTransfers(new_stats[type], transfers_players[league], "new/" + league + '/' + type + '.csv')
